airticket/airticket/src/main/java/com/example/airticket/controller/LocationController.java
package com.example.airticket.controller;

import com.example.airticket.entity.Location;
import com.example.airticket.service.LocationServiceImpl;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/location-api")
public class LocationController {
    private final LocationServiceImpl locationService;
    @Autowired
    public LocationController(LocationServiceImpl locationService) {
        this.locationService = locationService;
    }

    @GetMapping("/location")
    public ResponseEntity<List<Location>> getLocationDetails() {
        List<Location> locations = locationService.getLocationDetails();
        if (locations.isEmpty()) {
            return ResponseEntity.noContent().build();
        } else {
            return ResponseEntity.ok(locations);
        }
    }
}